【実装】:環境の切り分けなしでfirestoreの接続確認
